Code Bot/requirements.txt
discord.py
audioop-lts